requirements.txt
# AI
anthropic

# 이미지 처리
Pillow

# 브라우저 자동화
playwright

# HTTP / 네이버 API
requests
httpx

# 폴더 감시
watchdog

# 스케줄러
apscheduler

# 웹 서버 (Module 3 UI)
fastapi
uvicorn[standard]
jinja2
python-multipart

# 설정 관리
python-dotenv
pydantic
pydantic-settings

# HTML 파싱 (Module 3 URL 스타일 분석)
beautifulsoup4
lxml

# 유틸리티
aiofiles
loguru
tenacity
